add confirm dialog

+++ app/src/main/java/dev/leonlatsch/photok/cgallery/ui/navigation/GalleryNavigator.kt
@@ -16,7 +16,9 @@
 
 package dev.leonlatsch.photok.cgallery.ui.navigation
 
+import android.content.Context
 import androidx.core.os.bundleOf
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.navigation.NavController
 import dev.leonlatsch.photok.R
@@ -26,6 +28,7 @@ import dev.leonlatsch.photok.gallery.ui.menu.ExportBottomSheetDialogFragment
 import dev.leonlatsch.photok.model.database.entity.Photo
 import dev.leonlatsch.photok.other.INTENT_PHOTO_UUID
 import dev.leonlatsch.photok.other.extensions.show
+import dev.leonlatsch.photok.uicomponnets.Dialogs
 import javax.inject.Inject
 
 class GalleryNavigator @Inject constructor() {
@@ -33,29 +36,55 @@ class GalleryNavigator @Inject constructor() {
     fun navigate(
         event: GalleryNavigationEvent,
         navController: NavController,
-        fragmentManager: FragmentManager
+        fragment: Fragment,
     ) {
         when (event) {
             is GalleryNavigationEvent.OpenPhoto -> navigateOpenPhoto(event.photoUUID, navController)
-            GalleryNavigationEvent.OpenImportMenu -> navigateOpenImportMenu(fragmentManager)
+            GalleryNavigationEvent.OpenImportMenu -> navigateOpenImportMenu(fragment.childFragmentManager)
             is GalleryNavigationEvent.StartDeleteDialog -> navigateStartDeleteDialog(
+                fragment.requireContext(),
                 event.photosToDelete,
-                fragmentManager
+                fragment.childFragmentManager
             )
 
             is GalleryNavigationEvent.StartExportDialog -> navigateStartExportDialog(
+                fragment.requireContext(),
                 event.photosToExport,
-                fragmentManager
+                fragment.childFragmentManager
             )
         }
     }
 
-    private fun navigateStartExportDialog(photos: List<Photo>, fragmentManager: FragmentManager) {
-        ExportBottomSheetDialogFragment(photos).show(fragmentManager)
+    private fun navigateStartExportDialog(
+        context: Context,
+        photos: List<Photo>,
+        fragmentManager: FragmentManager
+    ) {
+        Dialogs.showConfirmDialog(
+            context,
+            String.format(
+                context.getString(R.string.export_are_you_sure),
+                photos.size
+            )
+        ) { _, _ -> // On positive button clicked
+            ExportBottomSheetDialogFragment(photos).show(fragmentManager)
+        }
     }
 
-    private fun navigateStartDeleteDialog(photos: List<Photo>, fragmentManager: FragmentManager) {
-        DeleteBottomSheetDialogFragment(photos).show(fragmentManager)
+    private fun navigateStartDeleteDialog(
+        context: Context,
+        photos: List<Photo>,
+        fragmentManager: FragmentManager
+    ) {
+        Dialogs.showConfirmDialog(
+            context,
+            String.format(
+                context.getString(R.string.delete_are_you_sure),
+                photos.size
+            )
+        ) { _, _ -> // On positive button clicked
+            DeleteBottomSheetDialogFragment(photos).show(fragmentManager)
+        }
     }
 
     private fun navigateOpenImportMenu(fragmentManager: FragmentManager) {
